requirements.txt
dj-database-url==0.4.0
Django==1.8.5
djangorestframework==3.2.4
gunicorn==19.4.5
Markdown==2.6.5
MySQL-python==1.2.5
